refactor : STEP 1 피드백 적용 안된 부분 적용
