+++ src/main/java/calculator/CalculatorMain.java
@@ -1,0 +1,11 @@
+package calculator;
+
+import calculator.domain.StringAddCalculator;
+
+public class CalculatorMain {
+
+  public static void main(String[] args) {
+    int result = StringAddCalculator.splitAndSum("1:2:3");
+    System.out.println(result);
+  }
+}
