+++ src/main/java/calculator/domain/StringAddCalculator.java
@@ -1,13 +1,6 @@
-package calculator;
-
-import calculator.domain.Formula;
+package calculator.domain;
 
 public class StringAddCalculator {
-
-  public static void main(String[] args) {
-    System.out.println(splitAndSum("1,2:3"));
-  }
-
   public static int splitAndSum(String userInput) {
     Formula calculator = Formula.createFormula(userInput);
     return calculator.summation();
