+++ src/test/java/calculator/StringAddCalculatorTest.java
@@ -3,6 +3,7 @@ package calculator;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import calculator.domain.StringAddCalculator;
 import org.junit.jupiter.api.Test;
 
 class StringAddCalculatorTest {
